REDESIGNED DOOR PREFABS, INTERACTIVE ANSWER UI SYSTEM, NUMBER GENERATOR
I made major changes to the code that governs the number generator however it is not done. Faith has extra requests for the problem questions so I am working on it.

I made major changes to the  Answer UI System, it is more interactive and more immediate- players no longer need to press TAB but rather can answer when they get close to a door.

I have changed the door animation to a door slide and will soon delete the need for a reactive Target. The player can simply press ENTER to open a door or signal a wrong answer.