No-Click, Swing Animation, Robot, Text Display of Question
I added a snippet of code to Nofer's fix. The text field would get deactivated when a player put in a wrong answer. It has been fixed.
I removed the 3rd person swing animation for now. I removed the from the teleport area because it would attack immediately. The text display nicely displays the integers in the question.